work/amocrm-dedup/dedup-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
